Removed output of hw3
Removed calls endScope(), printID() and printStructType(), because as stated in hw5, we should
remove outputs of hw3.
Error reports are still called, because I think it may hepl us in debugging, and inputs in hw5
won't contain errors anyway (according to what written in the FAQ).

Removed output of hw3

Removed calls endScope(), printID() and printStructType(), because as stated in hw5, we should
remove outputs of hw3.
Error reports are still called, because I think it may hepl us in debugging, and inputs in hw5
won't contain errors anyway (according to what written in the FAQ).